Updates to push geth log to websoxket in the live time. Updates to Use custom JsonDeserializer. Fixed Typo
